+++ source/extensions/filters/network/http_connection_manager/config.cc
@@ -35,8 +35,7 @@ namespace HttpConnectionManager {
 SINGLETON_MANAGER_REGISTRATION(date_provider);
 SINGLETON_MANAGER_REGISTRATION(route_config_provider_manager);
 
-Server::Configuration::NetworkFilterFactoryCb
-HttpConnectionManagerFilterConfigFactory::createFilter(
+Network::FilterFactoryCb HttpConnectionManagerFilterConfigFactory::createFilter(
     const envoy::config::filter::network::http_connection_manager::v2::HttpConnectionManager&
         proto_config,
     Server::Configuration::FactoryContext& context) {
@@ -50,9 +49,7 @@ HttpConnectionManagerFilterConfigFactory::createFilter(
   std::shared_ptr<Router::RouteConfigProviderManager> route_config_provider_manager =
       context.singletonManager().getTyped<Router::RouteConfigProviderManager>(
           SINGLETON_MANAGER_REGISTERED_NAME(route_config_provider_manager), [&context] {
-            return std::make_shared<Router::RouteConfigProviderManagerImpl>(
-                context.runtime(), context.dispatcher(), context.random(), context.localInfo(),
-                context.threadLocal(), context.admin());
+            return std::make_shared<Router::RouteConfigProviderManagerImpl>(context.admin());
           });
 
   std::shared_ptr<HttpConnectionManagerConfig> filter_config(new HttpConnectionManagerConfig(
@@ -69,16 +66,14 @@ HttpConnectionManagerFilterConfigFactory::createFilter(
   };
 }
 
-Server::Configuration::NetworkFilterFactoryCb
-HttpConnectionManagerFilterConfigFactory::createFilterFactory(
+Network::FilterFactoryCb HttpConnectionManagerFilterConfigFactory::createFilterFactory(
     const Json::Object& json_config, Server::Configuration::FactoryContext& context) {
   envoy::config::filter::network::http_connection_manager::v2::HttpConnectionManager proto_config;
   Config::FilterJson::translateHttpConnectionManager(json_config, proto_config);
   return createFilter(proto_config, context);
 }
 
-Server::Configuration::NetworkFilterFactoryCb
-HttpConnectionManagerFilterConfigFactory::createFilterFactoryFromProto(
+Network::FilterFactoryCb HttpConnectionManagerFilterConfigFactory::createFilterFactoryFromProto(
     const Protobuf::Message& proto_config, Server::Configuration::FactoryContext& context) {
   return createFilter(
       MessageUtil::downcastAndValidate<const envoy::config::filter::network::
@@ -133,9 +128,8 @@ HttpConnectionManagerConfig::HttpConnectionManagerConfig(
                                                                          context_.listenerScope())),
       proxy_100_continue_(config.proxy_100_continue()) {
 
-  route_config_provider_ = Router::RouteConfigProviderUtil::create(
-      config, context_.runtime(), context_.clusterManager(), context_.scope(), stats_prefix_,
-      context_.initManager(), route_config_provider_manager_);
+  route_config_provider_ = Router::RouteConfigProviderUtil::create(config, context_, stats_prefix_,
+                                                                   route_config_provider_manager_);
 
   switch (config.forward_client_cert_details()) {
   case envoy::config::filter::network::http_connection_manager::v2::HttpConnectionManager::SANITIZE:
@@ -255,7 +249,7 @@ HttpConnectionManagerConfig::HttpConnectionManagerConfig(
     auto& factory =
         Config::Utility::getAndCheckFactory<Server::Configuration::NamedHttpFilterConfigFactory>(
             string_name);
-    Server::Configuration::HttpFilterFactoryCb callback;
+    Http::FilterFactoryCb callback;
     if (filter_config->getBoolean("deprecated_v1", false)) {
       callback = factory.createFilterFactory(*filter_config->getObject("value", true),
                                              stats_prefix_, context);
@@ -294,7 +288,7 @@ HttpConnectionManagerConfig::createCodec(Network::Connection& connection,
 }
 
 void HttpConnectionManagerConfig::createFilterChain(Http::FilterChainFactoryCallbacks& callbacks) {
-  for (const Server::Configuration::HttpFilterFactoryCb& factory : filter_factories_) {
+  for (const Http::FilterFactoryCb& factory : filter_factories_) {
     factory(callbacks);
   }
 }