+++ source/server/lds_api.cc
@@ -40,7 +40,7 @@ void LdsApi::initialize(std::function<void()> callback) {
   subscription_->start({}, *this);
 }
 
-void LdsApi::onConfigUpdate(const ResourceVector& resources) {
+void LdsApi::onConfigUpdate(const ResourceVector& resources, const std::string& version_info) {
   cm_.adsMux().pause(Config::TypeUrl::get().RouteConfiguration);
   Cleanup rds_resume([this] { cm_.adsMux().resume(Config::TypeUrl::get().RouteConfiguration); });
   for (const auto& listener : resources) {
@@ -74,6 +74,7 @@ void LdsApi::onConfigUpdate(const ResourceVector& resources) {
     }
   }
 
+  version_info_ = version_info;
   runInitializeCallbackIfAny();
 }