+++ source/server/lds_subscription.h
@@ -42,15 +42,12 @@ private:
     NOT_IMPLEMENTED;
   }
 
-  const std::string versionInfo() const override { return version_info_; }
-
   // Http::RestApiFetcher
   void createRequest(Http::Message& request) override;
   void parseResponse(const Http::Message& response) override;
   void onFetchComplete() override;
   void onFetchFailure(const EnvoyException* e) override;
 
-  std::string version_info_;
   const LocalInfo::LocalInfo& local_info_;
   Config::SubscriptionCallbacks<envoy::api::v2::Listener>* callbacks_ = nullptr;
   Config::SubscriptionStats stats_;